codeforces_task_parser/requirements.txt
python-dotenv==0.19.0
psycopg2-binary==2.9.5
aiohttp~=3.8.4
tqdm~=4.65.0